+++ op_tests/op_benchmarks/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2024, Advanced Micro Devices, Inc. All rights reserved.
+from .triton.utils import *  # noqa: F403
