+++ op_tests/op_benchmarks/triton/utils/benchmark_utils.py
@@ -1,13 +1,12 @@
-import os
 import json
+import os
+import re
+import sys
+import tempfile
+import time
+
 import torch
 import triton.language as tl
-import sys
-import time
-import os
-import tempfile
-import re
-
 
 # Base directory where configs are located
 BASE_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../"))
