+++ op_tests/op_benchmarks/triton/utils/moe/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2024, Advanced Micro Devices, Inc. All rights reserved.
+from .utils import *  # noqa: F403
