+++ op_tests/op_benchmarks/triton/utils/moe/configs/device_name=AMD_Instinct_MI300X,dtype=int4_w4a16.json
