+++ op_tests/op_benchmarks/triton/utils/moe/configs/device_name=AMD_Instinct_MI300X,dtype=int8_w8a8.json
