+++ op_tests/triton_tests/utils/moe_align_block_size_ref.py
@@ -1,0 +1,118 @@
+import torch
+
+
+def _torch_moe_align_block_size(
+    topk_ids: torch.Tensor,
+    num_experts: int,
+    top_k: int,
+    block_size: int,
+    sorted_token_ids: torch.Tensor,
+    expert_ids: torch.Tensor,
+    num_tokens_post_pad: torch.Tensor,
+) -> None:
+    M, top_k = topk_ids.shape
+
+    expert_to_tokens = [[] for _ in range(num_experts)]
+    # For each token, for each selected expert, we append (token_id, expert)
+    for token_id in range(M):
+        for j in range(top_k):
+            e_id = topk_ids[token_id, j].item()
+            expert_to_tokens[e_id].append(token_id * top_k + j)
+
+    # Reorder tokens block by block, padding if needed
+    reordered_token_ids = []
+    reordered_expert_ids = []
+
+    tot_num_tokens = topk_ids.numel()
+    for e_id in range(num_experts):
+        tokens_for_expert = expert_to_tokens[e_id]
+        num_tokens = len(tokens_for_expert)
+
+        n_blocks = (num_tokens + block_size - 1) // block_size
+        # If not a multiple of block_size, pad up to the next multiple
+        padded_size = n_blocks * block_size
+
+        # Reorder all actual tokens for expert e_id
+        reordered_token_ids.extend(tokens_for_expert)
+        # reordered_expert_ids.extend([e_id]*num_tokens)
+        reordered_expert_ids.extend([e_id] * n_blocks)
+
+        # Pad with dummy token_id = -1 (or any sentinel), if needed
+        if padded_size > num_tokens:
+            pad_count = padded_size - num_tokens
+            reordered_token_ids.extend([tot_num_tokens] * pad_count)
+
+    token_length = len(reordered_token_ids)
+    expert_length = len(reordered_expert_ids)
+
+    sorted_token_ids[:token_length] = torch.tensor(
+        reordered_token_ids,
+        dtype=sorted_token_ids.dtype,
+        device=sorted_token_ids.device,
+    )
+    expert_ids[:expert_length] = torch.tensor(
+        reordered_expert_ids, dtype=expert_ids.dtype, device=expert_ids.device
+    )
+
+    # Fill remainder with -1 if these arrays are bigger than total_length
+    if token_length < sorted_token_ids.numel():
+        sorted_token_ids[token_length:] = tot_num_tokens
+    if expert_length < expert_ids.numel():
+        expert_ids[expert_length:] = tot_num_tokens
+
+    num_tokens_post_pad.fill_(token_length)
+
+
+def torch_moe_align_block_size(
+    topk_ids: torch.Tensor,  # [num_tkns, num_experts]
+    num_experts: int,
+    block_size: int,
+):
+    """
+    Aligns the token distribution across experts to be compatible with block size for matrix multiplication.
+
+    Parameters:
+    - topk_ids: A tensor of shape [total_tokens, top_k] representing the top-k expert indices for each token.
+    - block_size: The block size used in block matrix multiplication.
+    - num_experts: The total number of experts.
+
+    Returns:
+    - sorted_token_ids: A tensor containing the sorted token indices according to their allocated expert.
+    - expert_ids: A tensor indicating the assigned expert index for each block.
+    - num_tokens_post_padded: The total number of tokens after padding, ensuring divisibility by block_size.
+
+    This function pads the number of tokens that each expert needs to process so that it is divisible by block_size.
+    Padding ensures that during block matrix multiplication, the dimensions align correctly.
+
+    Example:
+    Given topk_ids = [[2, 3, 4], [1, 2, 4], [1, 3, 4], [1, 2, 3]], block_size = 4, and num_experts = 4:
+    - We initially have 12 tokens (after repeating 'top_k' times) and 4 experts, with each expert needing to process 3 tokens.
+    - As block_size is 4, we pad 1 token for each expert.
+    - First, flatten topk_ids to [2, 3, 4, 1, 2, 4, 1, 3, 4, 1, 2, 3].
+    - Then append padding tokens [12, 12, 12, 12] for each block.
+    - After sorting by expert index, we obtain token_ids [3, 6, 9, 12, 0, 4, 10, 12, 1, 7, 11, 12, 2, 5, 8, 12].
+        Tokens 12 are non-existent (padding) and are ignored in the subsequent matrix multiplication.
+    - The padding ensures that the total number of tokens is now divisible by block_size for proper block matrix operations.
+    """
+    top_k = topk_ids.shape[1]
+    max_num_tokens_padded = topk_ids.numel() + num_experts * (block_size - 1)
+    sorted_ids = torch.empty(
+        (max_num_tokens_padded,), dtype=torch.int32, device=topk_ids.device
+    )
+    max_num_m_blocks = (max_num_tokens_padded + block_size - 1) // block_size
+    expert_ids = torch.empty(
+        (max_num_m_blocks), dtype=torch.int32, device=topk_ids.device
+    )
+    sorted_ids.fill_(topk_ids.numel())
+    num_tokens_post_pad = torch.empty((1), dtype=torch.int32, device=topk_ids.device)
+    _torch_moe_align_block_size(
+        topk_ids,
+        num_experts,
+        top_k,
+        block_size,
+        sorted_ids,
+        expert_ids,
+        num_tokens_post_pad,
+    )
+
+    return sorted_ids, expert_ids, num_tokens_post_pad
